ver1_10/01/2020 Pagination in progress...

+++ eshop/src/main/java/com/soft/dao/EshopDAO.java
@@ -36,6 +36,6 @@ public interface EshopDAO {
          public List<BaseEntity> readEntityListByName(String entityName, Integer startRecord, Integer numOfRecords);
          
           //Method to get total number of records in DB table.     
-          public Long getTableSizeByTableName(String tableName);
+          public Integer getTableSizeByTableName(String tableName);
 }         
 //------------------------------------------------------------------------------
